requirements.txt
# Core Dependencies - Actually Used in Project
# ==============================================

# HTTP requests & web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Database
sqlalchemy>=2.0.23

# FastAPI and API dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6

# Configuration & validation
python-dotenv>=1.0.0
pydantic>=2.5.2

# Standard library enhancements
python-dateutil>=2.8.2

